test(cli): extract topics play test
Move the runtime-free game play topics coverage out of the play monolith into a focused suite and include that suite in the canonical play test command.

Proof boundary: test-only modularization. This layer changes no runtime direct-control behavior, play-thread flow, fake tuner ownership, oRPC transport, or relationship authority policy.
